- obtaion the dbtypes from the msyql dtabase. - add test for that. - start with a clean version of the knn algorithm and test for that.

+++ BlueFinderRS/src/knn/InstanceComparator.java
@@ -8,7 +8,7 @@ public class InstanceComparator implements Comparator<Instance>{
 	public int compare(Instance o1, Instance o2) {
 		Instance inst1 = (Instance) o1;
 		Instance inst2 = (Instance) o2;
-		int value = Float.compare(inst1.getDistance(), inst2.getDistance());
+		int value = Double.compare(inst1.getDistance(), inst2.getDistance());
 		if(value==0){
 			return inst1.getResource().compareTo(inst2.getResource());
 		}else{
